// src/types/PersonalDetails.types.ts
export interface IPersonalDetails {
    jobType: string;
    cnicNumber: string;
    feeDepositDate: Date;
    district: string;
    transcriptIssueDate: Date;
    applicantName: string;
    fatherName: string;
    domicile: string;
    dateOfBirth: Date;
    postalAddress: string;
    mobileNumber: string;
    emailAddress: string;
    hasDisability: boolean;
  }
  